--- src/freeFunctions/game/moveGunship.cpp
@@ -1,6 +1,0 @@
-#include "../../../include/models/iGunship.hpp"
-
-void moveGunship(IGunship &gunship, float x)
-{
-  gunship.move(x);
-}
