--- src/freeFunctions/game/moveGunshipLaser.cpp
@@ -1,6 +1,0 @@
-#include "../../../include/models/iLaser.hpp"
-
-void moveGunshipLaser(ILaser &gunshipLaser)
-{
-  gunshipLaser.move();
-}
