--- src/freeFunctions/game/moveMetroidLasers.cpp
@@ -1,10 +1,0 @@
-#include <vector>
-#include "../../../include/models/iLaser.hpp"
-
-void moveMetroidLasers(const std::vector<ILaser *> &metroidLasers)
-{
-  for (auto metroidLaser : metroidLasers)
-  {
-    metroidLaser->move();
-  }
-}
