+++ src/freeFunctions/game/moveObjects.cpp
@@ -1,7 +1,20 @@
 #include <vector>
+#include "../../../include/models/iGunship.hpp"
+#include "../../../include/models/iLaser.hpp"
 #include "../../../include/models/iMetroid.hpp"
+#include "../../../include/models/iRidley.hpp"
 #include "../../../include/wrappers/iClock.hpp"
 #include "../../../include/wrappers/iSound.hpp"
+
+void moveGunship(IGunship &gunship, float x)
+{
+  gunship.move(x);
+}
+
+void moveGunshipLaser(ILaser &gunshipLaser)
+{
+  gunshipLaser.move();
+}
 
 void moveMetroids(const std::vector<std::vector<IMetroid *>> &metroids, IClock &clock, int &interval, int &step, std::vector<ISound *> &sounds, int &soundCounter)
 {
@@ -27,4 +40,17 @@ void moveMetroids(const std::vector<std::vector<IMetroid *>> &metroids, IClock &
     sounds[soundCounter % 4]->play();
     soundCounter++;
   }
+}
+
+void moveMetroidLasers(const std::vector<ILaser *> &metroidLasers)
+{
+  for (auto metroidLaser : metroidLasers)
+  {
+    metroidLaser->move();
+  }
+}
+
+void moveRidley(IRidley &ridley)
+{
+  ridley.move();
 }
