--- src/freeFunctions/game/moveRidley.cpp
@@ -1,6 +1,0 @@
-#include "../../../include/models/iRidley.hpp"
-
-void moveRidley(IRidley &ridley)
-{
-  ridley.move();
-}
