+++ tests/freeFunctions/game/moveObjects/moveRidley.test.cpp
@@ -1,5 +1,5 @@
-#include "../../../include/game.hpp"
-#include "../../mockModels/mockRidley.hpp"
+#include "../../../../include/game.hpp"
+#include "../../../mockModels/mockRidley.hpp"
 
 TEST(moveRidley, callsMoveOnRidley)
 {
